Convert the DPU tmfifo_net0 mac into a variable - rshim_mac
rshim_mac can be specified as the complete 6-byte string ("AA:BB:CC:DD:EE:FF")
or as "random". If rshim_mac is specified as "random", a random mac is
generated.

If rshim_mac is not defined it is computed based on the rshim_num -
Tilera OUI-"00:1a:ca:" + "ff:ff"" + (3+(rshim_num*2)).

Signed-off-by: Anuradha Karuppiah <anuradhak@nvidia.com>